+++ src/main/java/Main.java
@@ -1,11 +1,11 @@
-import core.GameController;
+import core.controller.GameController;
 
 public class Main {
 
     public static void main(String[] args) {
 
         GameController gameController = GameController.getInstance();
-
+        gameController.start();
     }
 
 }
