--- src/main/java/core/GameController.java
@@ -1,17 +1,0 @@
-package core;
-
-public class GameController {
-
-
-
-
-    private static GameController gameController = new GameController();
-
-    public static GameController getInstance() {
-        return gameController;
-    }
-
-    private GameController() {
-
-    }
-}
